+++ eemeter/transform.py
@@ -293,14 +293,14 @@ def get_baseline_data(
     start_inf = False
     if start is None:
         # py datetime min/max are out of range of pd.Timestamp min/max
-        start_target = pytz.UTC.localize(pd.Timestamp.min)
+        start_target = pytz.UTC.localize(pd.Timestamp.min) + timedelta(days=1)
         start_inf = True
     else:
         start_target = start
 
     end_inf = False
     if end is None:
-        end_limit = pytz.UTC.localize(pd.Timestamp.max)
+        end_limit = pytz.UTC.localize(pd.Timestamp.max) - timedelta(days=1)
         end_inf = True
     else:
         end_limit = end
@@ -449,14 +449,14 @@ def get_reporting_data(
     start_inf = False
     if start is None:
         # py datetime min/max are out of range of pd.Timestamp min/max
-        start_limit = pytz.UTC.localize(pd.Timestamp.min)
+        start_limit = pytz.UTC.localize(pd.Timestamp.min) + timedelta(days=1)
         start_inf = True
     else:
         start_limit = start
 
     end_inf = False
     if end is None:
-        end_target = pytz.UTC.localize(pd.Timestamp.max)
+        end_target = pytz.UTC.localize(pd.Timestamp.max) - timedelta(days=1)
         end_inf = True
     else:
         end_target = end