+++ tests/test_transform.py
@@ -168,6 +168,14 @@ def test_get_baseline_data(il_electricity_cdd_hdd_hourly):
     assert len(warnings) == 0
 
 
+def test_get_baseline_data_with_timezones(il_electricity_cdd_hdd_hourly):
+    meter_data = il_electricity_cdd_hdd_hourly["meter_data"]
+    baseline_data, warnings = get_baseline_data(meter_data.tz_convert('America/New_York'))
+    assert len(warnings) == 0
+    baseline_data, warnings = get_baseline_data(meter_data.tz_convert('Australia/Sydney'))
+    assert len(warnings) == 0
+
+
 def test_get_baseline_data_with_end(il_electricity_cdd_hdd_hourly):
     meter_data = il_electricity_cdd_hdd_hourly["meter_data"]
     blackout_start_date = il_electricity_cdd_hdd_hourly["blackout_start_date"]
@@ -391,6 +399,14 @@ def test_get_reporting_data(il_electricity_cdd_hdd_hourly):
     meter_data = il_electricity_cdd_hdd_hourly["meter_data"]
     reporting_data, warnings = get_reporting_data(meter_data)
     assert meter_data.shape == reporting_data.shape == (19417, 1)
+    assert len(warnings) == 0
+
+
+def test_get_reporting_data_with_timezones(il_electricity_cdd_hdd_hourly):
+    meter_data = il_electricity_cdd_hdd_hourly["meter_data"]
+    reporting_data, warnings = get_reporting_data(meter_data.tz_convert('America/New_York'))
+    assert len(warnings) == 0
+    reporting_data, warnings = get_reporting_data(meter_data.tz_convert('Australia/Sydney'))
     assert len(warnings) == 0